requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9
python-dotenv>=1.0
requests>=2.31
djangorestframework>=3.15
